requirements.txt
Flask
gunicorn
pandas
numpy
sklearn
dill
six
-e .
PyYAML
ipykernel
requests
ipywidgets